+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Adjusts in settings with options always_online, read_messages and read_status
 * Fixed send webhook for event CALL
+* Create instance with settings
 
 # 1.4.2 (2023-07-24 20:52)